+++ apps/evals/src/tasks/chapter-lessons/task.ts
@@ -1,4 +1,4 @@
-import { generateChapterLessons } from "@zoonk/ai/chapter-lessons/generate";
+import { generateChapterLessons } from "@zoonk/ai/tasks/chapters/lessons";
 import type { Task } from "@/lib/types";
 import { TEST_CASES } from "./test-cases";
