+++ packages/ai/package.json
@@ -11,19 +11,19 @@
     "typescript": "^5"
   },
   "exports": {
-    "./activity-background/generate": "./src/activity-background/generate-activity-background.ts",
-    "./activity-explanation-quiz/generate": "./src/activity-explanation-quiz/generate-activity-explanation-quiz.ts",
-    "./activity-explanation/generate": "./src/activity-explanation/generate-activity-explanation.ts",
-    "./alternative-titles/generate": "./src/alternative-titles/generate-alternative-titles.ts",
-    "./chapter-lessons/generate": "./src/chapter-lessons/generate-chapter-lessons.ts",
-    "./course-categories/generate": "./src/course-categories/generate-course-categories.ts",
-    "./course-chapters/generate": "./src/course-chapters/generate-course-chapters.ts",
-    "./course-description/generate": "./src/course-description/generate-course-description.ts",
-    "./course-suggestions/generate": "./src/course-suggestions/generate-course-suggestions.ts",
-    "./course-thumbnail/generate": "./src/course-thumbnail/generate-course-thumbnail.ts",
     "./gateway": "./src/gateway.ts",
-    "./lesson-activities/generate": "./src/lesson-activities/generate-lesson-activities.ts",
-    "./lesson-kind/generate": "./src/lesson-kind/generate-lesson-kind.ts",
+    "./tasks/activities/core/background": "./src/tasks/activities/core/activity-background.ts",
+    "./tasks/activities/core/explanation": "./src/tasks/activities/core/activity-explanation.ts",
+    "./tasks/activities/core/explanation-quiz": "./src/tasks/activities/core/activity-explanation-quiz.ts",
+    "./tasks/chapters/lessons": "./src/tasks/chapters/chapter-lessons.ts",
+    "./tasks/courses/alternative-titles": "./src/tasks/courses/alternative-titles.ts",
+    "./tasks/courses/categories": "./src/tasks/courses/course-categories.ts",
+    "./tasks/courses/chapters": "./src/tasks/courses/course-chapters.ts",
+    "./tasks/courses/description": "./src/tasks/courses/course-description.ts",
+    "./tasks/courses/suggestions": "./src/tasks/courses/course-suggestions.ts",
+    "./tasks/courses/thumbnail": "./src/tasks/courses/course-thumbnail.ts",
+    "./tasks/lessons/activities": "./src/tasks/lessons/lesson-activities.ts",
+    "./tasks/lessons/kind": "./src/tasks/lessons/lesson-kind.ts",
     "./types": "./src/types.ts"
   },
   "name": "@zoonk/ai",
