+++ packages/ai/src/tasks/activities/_tools/fill-blank.prompt.md
