+++ packages/ai/src/tasks/activities/_tools/multiple-choice.prompt.md
