+++ packages/ai/src/tasks/activities/_tools/select-image.prompt.md
