+++ packages/ai/src/tasks/activities/core/activity-background.prompt.md
