+++ packages/ai/src/tasks/activities/core/activity-explanation-quiz.prompt.md
