+++ packages/ai/src/tasks/activities/core/activity-explanation.prompt.md
