+++ packages/ai/src/tasks/chapters/chapter-lessons.prompt.md
