+++ packages/ai/src/tasks/chapters/chapter-lessons.ts
