+++ packages/ai/src/tasks/courses/alternative-titles.prompt.md
