+++ packages/ai/src/tasks/courses/alternative-titles.ts
