+++ packages/ai/src/tasks/courses/course-categories.prompt.md
