+++ packages/ai/src/tasks/courses/course-categories.ts
