+++ packages/ai/src/tasks/courses/course-chapters.prompt.md
