+++ packages/ai/src/tasks/courses/course-chapters.ts
