+++ packages/ai/src/tasks/courses/course-description.prompt.md
