+++ packages/ai/src/tasks/courses/course-description.ts
