+++ packages/ai/src/tasks/courses/course-suggestions.prompt.md
