+++ packages/ai/src/tasks/courses/course-suggestions.ts
