+++ packages/ai/src/tasks/courses/course-thumbnail.ts
