+++ packages/ai/src/tasks/lessons/lesson-activities.prompt.md
