+++ packages/ai/src/tasks/lessons/lesson-activities.ts
