+++ packages/ai/src/tasks/lessons/lesson-kind.prompt.md
