+++ packages/ai/src/tasks/lessons/lesson-kind.ts
